+++ docs/CHANGELOG.md
@@ -8,6 +8,12 @@ The format is based on [Keep a Changelog][kac] and this project adheres to
 [kac]: https://keepachangelog.com/en/1.0.0/
 [semver]: https://semver.org/
 
+## [Unreleased]
+
+### Fixed
+
+* improved error trace for some broken cases (#279)
+
 ## [1.5.0] - 2021-10-22
 
 ### Added